Modify : model 피드백 수정

+++ cart/models.py
@@ -11,7 +11,7 @@ class Cart(models.Model):
 
 class Coupon(models.Model):
     name          = models.CharField(max_length=20)
-    discount_cost = models.IntegerField(null=True)
+    discount_cost = models.DecimalField(max_digits=10, decimal_places=2, null=True)
     discount_rate = models.DecimalField(max_digits=3, decimal_places=2, null=True)
 
     class Meta:
