+++ product/models.py
@@ -71,7 +71,7 @@ class Stock(models.Model):
 
 class Product(models.Model):
     title           = models.CharField(max_length=45)
-    price           = models.IntegerField()
+    price           = models.DecimalField(max_digits=10, decimal_places=2)
     model_name      = models.CharField(max_length=45)
     description     = models.TextField()
     discount_rate   = models.DecimalField(max_digits=3, decimal_places=2, null= True)
